Factor out adding courses and creating semesters
